Change tenant_static to frontend_static. This templatetag performs a mapping to cache-busting filenames and checks if a tenant-specific file exists, and returns that. If not it returns the normal path.
